ATO-692: Adds JWK handler
Adds a new route handler class for turning Public Keys into JWKs format.
I've added a cache control header of 1 Day as this seems to be done in
our other JWKs endpoints and could be useful for testing if/when we start
respecting this header. I've also added a content type as spark didn't
seem to render the JWK nicely without it.