ATO-1329: Fix claim format in request object
As part of other work[1], we're adding validation to the
claims included in request objects. This requires us
to put the claims nested under the 'userinfo' property
which the code is not currently doing. This corrects that
by attatching the claims as userinfo claims when making a request
object authorize JAR

[1]- https://github.com/govuk-one-login/authentication-api/pull/5721